--- README.md
@@ -1,1 +1,0 @@
-# Geospatial-Analysis-of-NexRad-and-GEOS-satellite
