--- great-expectations/great_expectations/.gitignore
@@ -1,1 +1,0 @@
-uncommitted/
