--- great-expectations/great_expectations/data/Spring 2023 - DAMG 7245 Group exercises - GEOS - Datasets.csv
@@ -1,13 +1,0 @@
-Group No,File name,Name,Full file name,Colab link,Name,Does your team's file link work?,Link of your next team works?
-1,OR_ABI-L2-ACMM1-M6_G18_s20230090504262_e20230090504319_c20230090505026.nc,Thejas,https://noaa-goes18.s3.amazonaws.com/ABI-L2-ACMM/2023/009/05/OR_ABI-L2-ACMM1-M6_G18_s20230090504262_e20230090504319_c20230090505026.nc,https://colab.research.google.com/drive/1_4P4gHRSfAYo4SrIufoPU-9iFvQCPhhi,Thejas,Yes,Yes
-2,OR_ABI-L2-ACTPM1-M6_G18_s20230090408262_e20230090408319_c20230090409174.nc,Meet/Ajin/Siddhi,https://noaa-goes18.s3.amazonaws.com/ABI-L2-ACTPM/2023/009/04/OR_ABI-L2-ACTPM1-M6_G18_s20230090408262_e20230090408319_c20230090409174.nc,https://colab.research.google.com/drive/18Crm5b56mSwIK0wOxlZpKAGWIoS_PiVv#scrollTo=6VZyku18JE-p,Ajin/Siddhi/Meet,Yes,Yes
-3,OR_ABI-L2-DSIM1-M6_G18_s20230110608251_e20230110608308_c20230110609126.nc,Raj/Mani,https://noaa-goes18.s3.amazonaws.com/ABI-L2-DSIM/2023/011/06/OR_ABI-L2-DSIM1-M6_G18_s20230110608251_e20230110608308_c20230110609126.nc,https://colab.research.google.com/drive/1c_lIvojuikRDBs-jnLRyCRWjMEiNZIzy?usp=sharing,"Raj, Mani",Yes,Yes
-4,OR_ABI-L2-ACHTM1-M6_G18_s20223560805242_e20223560805300_c20223560806526.nc,Rishabh,https://noaa-goes18.s3.amazonaws.com/ABI-L2-ACHTM/2022/356/08/OR_ABI-L2-ACHTM1-M6_G18_s20223560805242_e20223560805300_c20223560806526.nc,https://colab.research.google.com/drive/1lLjB3csqzb_JQ9y4TtglxOjQ5E1H5tHc,All,Yes,Yes
-5,OR_ABI-L2-BRFF-M6_G18_s20223150230207_e20223150239515_c20223150241087.nc,Charan,https://noaa-goes18.s3.amazonaws.com/ABI-L2-BRFF/2022/315/02/OR_ABI-L2-BRFF-M6_G18_s20223150230207_e20223150239515_c20223150241087.nc,https://colab.research.google.com/drive/1K3__EwbfwD2yGOsW-yqjwkBGW8qIIp9i?usp=sharing,"Rishabh Jain, Charan, Krishica",Yes,Yes
-6,OR_ABI-L2-ADPM2-M6_G18_s20230061310557_e20230061311015_c20230061311402.nc,Midhun,https://noaa-goes18.s3.amazonaws.com/ABI-L2-ADPM/2023/006/13/OR_ABI-L2-ADPM2-M6_G18_s20230061310557_e20230061311015_c20230061311402.nc,https://colab.research.google.com/drive/1yaJ8QpQ5vDJvMnRcNEB9YxZT3hTQsRe9?usp=sharing,All,Yes,Yes
-7,OR_ABI-L1b-RadM1-M6C01_G18_s20230030201252_e20230030201311_c20230030201340.nc,Shubham,https://noaa-goes18.s3.amazonaws.com/ABI-L1b-RadM/2023/003/02/OR_ABI-L1b-RadM1-M6C01_G18_s20230030201252_e20230030201311_c20230030201340.nc,https://colab.research.google.com/drive/11_8oFk6sGT7Vw7JBHxfWhm94KLb9kq3s?usp=sharing,Shubham,Yes,Yes
-8,OR_ABI-L2-ACHTF-M6_G18_s20223532240210_e20223532249518_c20223532252164.nc,Vraj,https://noaa-goes18.s3.amazonaws.com/ABI-L2-ACHTF/2022/353/22/OR_ABI-L2-ACHTF-M6_G18_s20223532240210_e20223532249518_c20223532252164.nc,https://colab.research.google.com/drive/1qw76y5t2R1nrVqJ2H2Rko6zjjeLqi0_H?usp=sharing,All,Yes,Yes
-9,OR_ABI-L2-DSRC-M6_G18_s20223180501179_e20223180503552_c20223180508262.nc,Lakshman,https://noaa-goes18.s3.amazonaws.com/ABI-L2-DSRC/2022/318/05/OR_ABI-L2-DSRC-M6_G18_s20223180501179_e20223180503552_c20223180508262.nc,https://colab.research.google.com/drive/1x_jiZ6shWwETsD0ISXyKhP8lbxldmx_q?usp=sharing,All,Yes,Yes
-10,OR_ABI-L2-DMWVM1-M6C08_G18_s20223552050271_e20223552050328_c20223552122197.nc,Snehashis,https://noaa-goes18.s3.amazonaws.com/ABI-L2-DMWVM/2022/355/20/OR_ABI-L2-DMWVM1-M6C08_G18_s20223552050271_e20223552050328_c20223552122197.nc,https://colab.research.google.com/drive/1yM-Eq7srEe8bq2HUSVFHd5M7ZFyqft9A#scrollTo=XLbaljgMOQAn,All,Yes,Yes
-11,OR_ABI-L2-ACMC-M6_G18_s20222800931164_e20222800933537_c20222800934574.nc,Aakash,https://noaa-goes18.s3.amazonaws.com/ABI-L2-ACMC/2022/280/09/OR_ABI-L2-ACMC-M6_G18_s20222800931164_e20222800933537_c20222800934574.nc,https://colab.research.google.com/drive/15KWiOU8uM4AjqXbt4RjAGxCC-ntHB7np?usp=sharing,Bhargavi,Yes,Yes
-12,OR_ABI-L2-DMWC-M6C07_G18_s20223510516174_e20223510518559_c20223510527449.nc,DEV,https://noaa-goes18.s3.amazonaws.com/ABI-L2-DMWC/2022/351/05/OR_ABI-L2-DMWC-M6C07_G18_s20223510516174_e20223510518559_c20223510527449.nc,https://colab.research.google.com/drive/1esO4QpUQba4Sr7X5YH6JwG1-NQhYVrMk?usp=sharing#scrollTo=DDDSVH0r6kiv,Harsh,Yes,Yes
