--- great-expectations/great_expectations/expectations/.ge_store_backend_id
@@ -1,1 +1,0 @@
-store_backend_id = 2d46882f-04a4-4d99-aebf-84b8f83f9295
