--- great-expectations/great_expectations/expectations/GOES_suite.json
@@ -1,54 +1,0 @@
-{
-  "data_asset_type": null,
-  "expectation_suite_name": "GOES_suite",
-  "expectations": [
-    {
-      "expectation_type": "expect_table_row_count_to_be_between",
-      "kwargs": {
-        "max_value": 12,
-        "min_value": 12
-      },
-      "meta": {
-        "profiler_details": {
-          "metric_configuration": {
-            "domain_kwargs": {},
-            "metric_name": "table.row_count",
-            "metric_value_kwargs": null
-          },
-          "num_batches": 1
-        }
-      }
-    },
-    {
-      "expectation_type": "expect_table_columns_to_match_set",
-      "kwargs": {
-        "column_set": [
-          "File name",
-          "Does your team's file link work?",
-          "Name.1",
-          "Colab link",
-          "Full file name",
-          "Link of your next team works?",
-          "Name",
-          "Group No"
-        ],
-        "exact_match": null
-      },
-      "meta": {
-        "profiler_details": {
-          "success_ratio": 1.0
-        }
-      }
-    }
-  ],
-  "ge_cloud_id": null,
-  "meta": {
-    "citations": [
-      {
-        "citation_date": "2023-02-10T04:28:59.663303Z",
-        "comment": "Created by effective Rule-Based Profiler of OnboardingDataAssistant with the configuration included.\n"
-      }
-    ],
-    "great_expectations_version": "0.15.47"
-  }
-}
