--- great-expectations/great_expectations/great_expectations.yml
@@ -1,135 +1,0 @@
-# Welcome to Great Expectations! Always know what to expect from your data.
-#
-# Here you can define datasources, batch kwargs generators, integrations and
-# more. This file is intended to be committed to your repo. For help with
-# configuration please:
-#   - Read our docs: https://docs.greatexpectations.io/docs/guides/connecting_to_your_data/connect_to_data_overview/#2-configure-your-datasource
-#   - Join our slack channel: http://greatexpectations.io/slack
-
-# config_version refers to the syntactic version of this config file, and is used in maintaining backwards compatibility
-# It is auto-generated and usually does not need to be changed.
-config_version: 3.0
-
-# Datasources tell Great Expectations where your data lives and how to get it.
-# You can use the CLI command `great_expectations datasource new` to help you
-# add a new datasource. Read more at https://docs.greatexpectations.io/docs/guides/connecting_to_your_data/connect_to_data_overview
-datasources:
-  goes_datasource:
-    module_name: great_expectations.datasource
-    execution_engine:
-      class_name: PandasExecutionEngine
-      module_name: great_expectations.execution_engine
-    class_name: Datasource
-    data_connectors:
-      default_inferred_data_connector_name:
-        default_regex:
-          group_names:
-            - data_asset_name
-          pattern: (.*)\.csv
-        module_name: great_expectations.datasource.data_connector
-        base_directory: ..\data
-        class_name: InferredAssetFilesystemDataConnector
-        name: default_inferred_data_connector_name
-      default_runtime_data_connector_name:
-        module_name: great_expectations.datasource.data_connector
-        assets:
-          my_runtime_asset_name:
-            module_name: great_expectations.datasource.data_connector.asset
-            class_name: Asset
-            batch_identifiers:
-              - runtime_batch_identifier_name
-        class_name: RuntimeDataConnector
-        name: default_runtime_data_connector_name
-  GOES_datasource:
-    module_name: great_expectations.datasource
-    execution_engine:
-      class_name: PandasExecutionEngine
-      module_name: great_expectations.execution_engine
-    class_name: Datasource
-    data_connectors:
-      default_inferred_data_connector_name:
-        default_regex:
-          group_names:
-            - data_asset_name
-          pattern: (.*)\.csv
-        module_name: great_expectations.datasource.data_connector
-        base_directory: data
-        class_name: InferredAssetFilesystemDataConnector
-      default_runtime_data_connector_name:
-        module_name: great_expectations.datasource.data_connector
-        assets:
-          my_runtime_asset_name:
-            module_name: great_expectations.datasource.data_connector.asset
-            class_name: Asset
-            batch_identifiers:
-              - runtime_batch_identifier_name
-        class_name: RuntimeDataConnector
-config_variables_file_path: uncommitted/config_variables.yml
-
-# The plugins_directory will be added to your python path for custom modules
-# used to override and extend Great Expectations.
-plugins_directory: plugins/
-
-stores:
-# Stores are configurable places to store things like Expectations, Validations
-# Data Docs, and more. These are for advanced users only - most users can simply
-# leave this section alone.
-#
-# Three stores are required: expectations, validations, and
-# evaluation_parameters, and must exist with a valid store entry. Additional
-# stores can be configured for uses such as data_docs, etc.
-  expectations_store:
-    class_name: ExpectationsStore
-    store_backend:
-      class_name: TupleFilesystemStoreBackend
-      base_directory: expectations/
-
-  validations_store:
-    class_name: ValidationsStore
-    store_backend:
-      class_name: TupleFilesystemStoreBackend
-      base_directory: uncommitted/validations/
-
-  evaluation_parameter_store:
-    class_name: EvaluationParameterStore
-  checkpoint_store:
-    class_name: CheckpointStore
-    store_backend:
-      class_name: TupleFilesystemStoreBackend
-      suppress_store_backend_id: true
-      base_directory: checkpoints/
-
-  profiler_store:
-    class_name: ProfilerStore
-    store_backend:
-      class_name: TupleFilesystemStoreBackend
-      suppress_store_backend_id: true
-      base_directory: profilers/
-
-expectations_store_name: expectations_store
-validations_store_name: validations_store
-evaluation_parameter_store_name: evaluation_parameter_store
-checkpoint_store_name: checkpoint_store
-
-data_docs_sites:
-  # Data Docs make it simple to visualize data quality in your project. These
-  # include Expectations, Validations & Profiles. The are built for all
-  # Datasources from JSON artifacts in the local repo including validations &
-  # profiles from the uncommitted directory. Read more at https://docs.greatexpectations.io/docs/terms/data_docs
-  local_site:
-    class_name: SiteBuilder
-    show_how_to_buttons: true
-    store_backend:
-      class_name: TupleFilesystemStoreBackend
-      base_directory: uncommitted/data_docs/local_site/
-    site_index_builder:
-      class_name: DefaultSiteIndexBuilder
-
-anonymous_usage_statistics:
-  enabled: true
-  data_context_id: 2d46882f-04a4-4d99-aebf-84b8f83f9295
-notebooks:
-include_rendered_content:
-  globally: false
-  expectation_suite: false
-  expectation_validation_result: false
