--- great-expectations/great_expectations/plugins/custom_data_docs/styles/data_docs_custom_styles.css
@@ -1,22 +1,0 @@
-/*index page*/
-.ge-index-page-site-name-title {}
-.ge-index-page-table-container {}
-.ge-index-page-table {}
-.ge-index-page-table-profiling-links-header {}
-.ge-index-page-table-expectations-links-header {}
-.ge-index-page-table-validations-links-header {}
-.ge-index-page-table-profiling-links-list {}
-.ge-index-page-table-profiling-links-item {}
-.ge-index-page-table-expectation-suite-link {}
-.ge-index-page-table-validation-links-list {}
-.ge-index-page-table-validation-links-item {}
-
-/*breadcrumbs*/
-.ge-breadcrumbs {}
-.ge-breadcrumbs-item {}
-
-/*navigation sidebar*/
-.ge-navigation-sidebar-container {}
-.ge-navigation-sidebar-content {}
-.ge-navigation-sidebar-title {}
-.ge-navigation-sidebar-link {}
