+++ requirements.txt
@@ -1,0 +1,64 @@
+aiobotocore==2.4.2
+aiohttp==3.8.3
+aioitertools==0.11.0
+aiosignal==1.3.1
+altair==4.2.2
+async-timeout==4.0.2
+attrs==22.2.0
+blinker==1.5
+boto3==1.26.66
+botocore==1.27.59
+cachetools==5.3.0
+certifi==2022.12.7
+charset-normalizer==2.1.1
+click==8.1.3
+decorator==5.1.1
+entrypoints==0.4
+frozenlist==1.3.3
+fsspec==2023.1.0
+gitdb==4.0.10
+GitPython==3.1.30
+idna==3.4
+importlib-metadata==6.0.0
+Jinja2==3.1.2
+jmespath==1.0.1
+jsonschema==4.17.3
+markdown-it-py==2.1.0
+MarkupSafe==2.1.2
+mdurl==0.1.2
+multidict==6.0.4
+numpy==1.24.2
+packaging==23.0
+pandas==1.5.3
+Pillow==9.4.0
+plotly==5.13.0
+protobuf==3.20.3
+pyarrow==11.0.0
+pydeck==0.8.0
+Pygments==2.14.0
+Pympler==1.0.1
+pyrsistent==0.19.3
+python-dateutil==2.8.2
+pytz==2022.7.1
+pytz-deprecation-shim==0.1.0.post0
+requests==2.28.2
+rich==13.3.1
+s3fs==2023.1.0
+s3transfer==0.6.0
+semver==2.13.0
+six==1.16.0
+smmap==5.0.0
+streamlit==1.17.0
+tenacity==8.2.0
+toml==0.10.2
+toolz==0.12.0
+tornado==6.2
+tqdm==4.64.1
+typing_extensions==4.4.0
+tzdata==2022.7
+tzlocal==4.2
+urllib3==1.26.14
+validators==0.20.0
+wrapt==1.14.1
+yarl==1.8.2
+zipp==3.12.1
