--- streamlit/.streamlit/secrets.toml
@@ -1,2 +1,0 @@
-AWS_ACCESS_KEY_ID = ""
-AWS_SECRET_ACCESS_KEY = ""
