--- streamlit/nexrad_loc.csv
@@ -1,157 +1,0 @@
-State,City,ICAO Location Identifier,Lat,Long
-PR,San Juan,TJUA,18.1155998,-66.0780644
-ME,Houlton,KCBW,46.0391944,-67.8066033
-ME,Gray/Portland,KGYX,43.8913555,-70.2565545
-VT,Burlington,KCXX,44.5109941,-73.166424
-MA,Boston,KBOX,41.9558919,-71.1369681
-NY,Albany,KENX,42.5865699,-74.0639877
-NY,Binghamton,KBGM,42.1997045,-75.9847015
-NY,Buffalo,KBUF,42.9488055,-78.7369108
-NY,Montague,KTYX,43.7556319,-75.6799918
-NY,New York City,KOKX,40.8655093,-72.8638548
-DE,Dover AFB,KDOX,38.8257651,-75.4400763
-PA,Philadelphia,KDIX,39.9470885,-74.4108027
-PA,Pittsburgh,KPBZ,40.5316842,-80.2179515
-PA,State College,KCCX,40.9228521,-78.0038738
-WV,Charleston,KRLX,38.3110763,-81.7229015
-VA,Norfolk/Richmond,KAKQ,36.9840475,-77.007342
-VA,Roanoke,KFCX,37.0242098,-80.2736664
-VA,Sterling,KLWX,38.9753957,-77.4778444
-NC,Morehead City,KMHX,34.7759313,-76.8762571
-NC,Raleigh/Durham,KRAX,35.6654967,-78.4897855
-NC,Wilmington,KLTX,33.9891631,-78.4291059
-SC,Charleston,KCLX,32.6554866,-81.0423124
-SC,Columbia,KCAE,33.9487579,-81.1184281
-SC,Greer,KGSP,34.8833435,-82.2200757
-GA,Atlanta,KFFC,33.3635771,-84.565866
-GA,Moody AFB,KVAX,30.8903853,-83.0019021
-GA,Robins AFB,KJGX,32.6755239,-83.3508575
-FL,Eglin AFB,KEVX,30.5649908,-85.921559
-FL,Jacksonville,KJAX,30.4846878,-81.7018917
-FL,Key West,KBYX,24.5974996,-81.7032355
-FL,Melbourne,KMLB,28.1131808,-80.6540988
-FL,Miami,KAMX,25.6111275,-80.412747
-FL,Tallahassee,KTLH,30.397568,-84.3289116
-FL,Tampa,KTBW,27.7054701,-82.40179
-AL,Birmingham,KBMX,33.1722806,-86.7698425
-AL,Fort Rucker,KEOX,31.4605622,-85.4592401
-AL,Huntsville,KHTX,34.930508,-86.0837388
-AL,Maxwell AFB,KMXX,32.5366608,-85.7897848
-AL,Mobile,KMOB,30.6795378,-88.2397816
-MS,Brandon/Jackson,KDGX,32.2797358,-89.9846309
-MS,Columbus AFB,KGWX,33.8967796,-88.3293915
-TN,Knoxville/Tri Cities,KMRX,36.168538,-83.401779
-TN,Memphis,KNQA,35.3447802,-89.8734534
-TN,Nashville,KOHX,36.2472389,-86.5625185
-KY,Fort Campbell,KHPX,36.7368894,-87.2854328
-KY,Jackson,KJKL,37.590762,-83.313039
-KY,Louisville,KLVX,37.9753058,-85.9438455
-KY,Paducah,KPAH,37.0683618,-88.7720257
-OH,Wilmington,KILN,39.42028,-83.82167
-OH,Cleveland,KCLE,41.4131875,-81.8597451
-MI,Detroit/Pontiac,KDTX,42.6999677,-83.471809
-MI,Gaylord,KAPX,44.907106,-84.719817
-MI,Grand Rapids,KGRR,42.893872,-85.5449206
-MI,Marquette,KMQT,46.5311443,-87.5487131
-IN,Owensville (Evansville),KVWX,38.2603901,-87.7246553
-IN,Indianapolis,KIND,39.7074962,-86.2803675
-IN,North Webster,KIWX,41.3586356,-85.7000488
-IL,Chicago,KLOT,41.6044264,-88.084361
-IL,Lincoln,KILX,40.150544,-89.336842
-WI,Green Bay,KGRB,44.4984644,-88.111124
-WI,La Crosse,KARX,43.822766,-91.1915767
-WI,Milwaukee,KMKX,42.9678286,-88.5506335
-MN,Duluth,KDLH,46.8368569,-92.2097433
-MN,Minneapolis/St. Paul,KMPX,44.8488029,-93.5654873
-IA,Davenport,KDVN,41.611556,-90.5809987
-IA,Des Moines,KDMX,41.7311788,-93.7229235
-MO,Kansas City,KEAX,38.8102231,-94.2644924
-MO,Springfield,KSGF,37.235223,-93.4006011
-MO,St. Louis,KLSX,38.6986863,-90.682877
-AR,Fort Smith,KSRX,35.2904423,-94.3619075
-AR,Little Rock,KLZK,34.8365261,-92.2621697
-LA,Fort Polk,KPOE,31.1556923,-92.9762596
-LA,Lake Charles,KLCH,30.125382,-93.2161188
-LA,New Orleans,KLIX,30.3367133,-89.8256618
-LA,Shreveport,KSHV,32.450813,-93.8412774
-TX,Amarillo,KAMA,35.2334827,-101.7092478
-TX,Austin/San Antonio,KEWX,29.7039802,-98.028506
-TX,Brownsville,KBRO,25.9159979,-97.4189526
-TX,Corpus Christi,KCRP,27.7840203,-97.511234
-TX,Dallas/Ft. Worth,KFWS,32.5730186,-97.3031911
-TX,Dyess AFB,KDYX,32.5386009,-99.2542863
-TX,El Paso,KEPZ,31.8731115,-106.697942
-TX,Fort Hood,KGRK,30.7217637,-97.3829627
-TX,Houston/Galveston,KHGX,29.4718835,-95.0788593
-TX,Laughlin AFB,KDFX,29.2730823,-100.2802312
-TX,Lubbock,KLBB,33.6541242,-101.814149
-TX,Midland/Odessa,KMAF,31.9433953,-102.1894383
-TX,San Angelo,KSJT,31.3712815,-100.4925227
-OK,Frederick,KFDR,34.3620014,-98.9766884
-OK,Oklahoma City,KTLX,35.3333873,-97.2778255
-OK,Norman (Testbed),No Level III Data,35.2358,-97.4622
-OK,Tulsa,KINX,36.1750977,-95.5642802
-OK,Vance AFB,KVNX,36.7406166,-98.1279409
-KS,Dodge City,KDDC,37.7608043,-99.9688053
-KS,Goodland,KGLD,39.3667737,-101.7004341
-KS,Topeka,KTWX,38.996998,-96.232618
-KS,Wichita,KICT,37.6545724,-97.4431461
-NE,Grand Island/Hastings,KUEX,40.320966,-98.4418559
-NE,North Platte,KLNX,41.9579623,-100.5759609
-NE,Omaha,KOAX,41.3202803,-96.3667971
-SD,Aberdeen,KABR,45.4558185,-98.4132046
-SD,Rapid City,KUDX,44.1248485,-102.8298157
-SD,Sioux Falls,KFSD,43.5877467,-96.7293674
-ND,Bismarck,KBIS,46.7709329,-100.7605532
-ND,Grand Forks (Mayville),KMVX,47.5279417,-97.3256654
-ND,Minot AFB,KMBX,48.39303,-100.8644378
-MT,Billings,KBLX,45.8537632,-108.6068165
-MT,Glasgow,KGGW,48.2064536,-106.6252971
-MT,Great Falls,KTFX,47.4595023,-111.3855368
-MT,Missoula,KMSX,47.0412971,-113.9864373
-WY,Cheyenne,KCYS,41.1519308,-104.8060325
-WY,Riverton,KRIW,43.0660779,-108.4773731
-CO,Denver,KFTG,39.7866156,-104.5458126
-CO,Grand Junction,KGJX,39.0619824,-108.2137012
-CO,Pueblo,KPUX,38.4595034,-104.1816223
-NM,Albuquerque,KABX,35.1497579,-106.8239576
-NM,Cannon AFB,KFDX,34.6341569,-103.6186427
-NM,Holloman AFB,KHDX,33.0768844,-106.1200923
-AZ,Flagstaff,KFSX,34.574449,-111.198367
-AZ,Phoenix,KIWA,33.289111,-111.6700092
-AZ,Tucson,KEMX,31.8937186,-110.6304306
-AZ,Yuma,KYUX,32.4953477,-114.6567214
-UT,Cedar City,KICX,37.59083,-112.86222
-UT,Salt Lake City,KMTX,41.2627795,-112.4480081
-ID,Boise,KCBX,43.4902104,-116.2360436
-ID,Pocatello/Idaho Falls,KSFX,43.1055967,-112.6860487
-NV,Elko,KLRX,40.7396933,-116.8025529
-NV,Las Vegas,KESX,35.7012894,-114.8918277
-NV,Reno,KRGX,39.7541931,-119.4620597
-CA,Beale AFB,KBBX,39.4956958,-121.6316557
-CA,Edwards AFB,KEYX,35.0979358,-117.5608832
-CA,Eureka,KBHX,40.4986955,-124.2918867
-CA,Los Angeles,KVTX,34.4116386,-119.1795641
-CA,Sacramento,KDAX,38.5011529,-121.6778487
-CA,San Diego,KNKX,32.9189891,-117.041814
-CA,San Francisco,KMUX,37.155152,-121.8984577
-CA,San Joaquin Valley,KHNX,36.3142088,-119.6320903
-CA,Santa Ana Mountains,KSOX,33.8176452,-117.6359743
-CA,Vandenberg AFB,KVBG,34.8383137,-120.3977805
-HI,Kauai,PHKI,21.8938762,-159.5524585
-HI,Kohala,PHKM,20.1254606,-155.778054
-HI,Molokai,PHMO,21.1327531,-157.1802807
-HI,South Shore,PHWA,19.0950155,-155.5688846
-OR,Medford,KMAX,42.0810766,-122.7173334
-OR,Pendleton,KPDT,45.6906118,-118.8529301
-OR,Portland,KRTX,45.7150308,-122.9650542
-WA,Langley Hill,KLGX,47.116806,-124.10625
-WA,Seattle/Tacoma,KATX,48.1945614,-122.4957508
-WA,Spokane,KOTX,47.6803744,-117.6267797
-AK,Bethel,PABC,60.791987,-161.876539
-AK,Fairbanks/Pedro Dome,PAPD,65.0351238,-147.5014222
-AK,Kenai,PAHG,60.725833,-151.351389
-AK,King Salmon,PAKC,58.6794558,-156.6293335
-AK,Middleton Island,PAIH,59.46194,-146.30111
-AK,Nome,PAEC,64.5114973,-165.2949071
-AK,Sitka/Biorka Island,PACG,56.85214,-135.552417
